[ update ] Upcoming tightening of the parser
https://github.com/idris-lang/Idris2/pull/1711 makes the parser
stricter and in exchange returns better error messages in some
cases that seem to trip beginners over often.

We are particularly hit by the fact that we can't have dangling
let / rewrite / lam (that last one is the big one) as the last
argument of an application.

The fix is easy: insert many, many `$`.